requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
apturl==0.5.2
attrs==25.3.0
blinker==1.9.0
Brlapi==0.8.3
cachetools==6.1.0
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.4.2
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
exceptiongroup==1.3.0
fastapi==0.115.13
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.3
h11==0.16.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
macaroonbakery==1.3.1
MarkupSafe==3.0.2
more-itertools==8.10.0
narwhals==1.43.1
netifaces==0.11.0
numpy==2.2.6
oauthlib==3.2.0
olefile==0.46
packaging==25.0
pandas==2.3.0
pexpect==4.8.0
Pillow==9.0.1
protobuf==6.31.1
ptyprocess==0.7.0
pyarrow==20.0.0
pycairo==1.20.1
pycups==2.0.1
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.1.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
referencing==0.36.2
reportlab==3.6.8
requests==2.32.4
rpds-py==0.25.1
SecretStorage==3.3.1
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlmodel==0.0.24
starlette==0.46.2
streamlit==1.46.0
systemd-python==234
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
uvicorn==0.34.3
wadllib==1.3.6
watchdog==6.0.0
xdg==5
xkit==0.0.0
zipp==1.0.0
